Make all Round related types explicit
