+++ src/test-factories/round.ts
@@ -1,6 +1,12 @@
 import { useRules } from '@/composables/useRules';
 
-import { useRoundsStore, type Scores } from '@/stores/rounds';
+import {
+	useRoundsStore,
+	type CurrentRound,
+	type CompletedRound,
+	type RoundPhase,
+	type Round
+} from '@/stores/rounds';
 
 import { generateId } from '@/utilities/id';
 
@@ -20,7 +26,7 @@ function createCurrentRound(playerIds: Id[], phase: RoundPhase = 'turns', winner
 	} satisfies CurrentRound;
 }
 
-function createCompletedRound(winnerId: Id, scores: Scores): CompletedRound {
+function createCompletedRound(winnerId: Id, scores: PlayerScoreMap): CompletedRound {
 	return {
 		id: generateId(),
 		isCurrentRound: false,
